refactor(kftray-kube): promote kube/client/ to crate root
