+++ crates/kftray-kube/src/client/error.rs
