+++ crates/kftray-kube/src/lib.rs
@@ -1,10 +1,11 @@
+pub mod client;
 pub mod dataplane_runtime;
 pub mod kube;
 pub mod port_forward;
 pub mod port_forward_error;
 pub mod registry;
 
-pub use kube::client::create_client_with_specific_context;
+pub use client::create_client_with_specific_context;
 pub use kube::models::{
     KubeContextInfo,
     KubeNamespaceInfo,
